refactor: move emoji encoding functions to notifier.py and remove utils.py
